Update patch set 4
Patch Set 4: Code-Review-1

I don't think UniquePtr is appropriate here.  At least not in the way it has been applied here.  It should be all-or-nothing to ensure that the storage is appropriately managed.

The problem is that we're passing storage allocated with malloc() to UniquePtr which will eventually use delete to free it.  malloc()'d memory should be free()'d.

Mixing malloc() and new is probably safe in Bionic but technically they could use different allocators and the behavior is undefined.

I'm better disposed to the earlier patch.  This should really just involve adding a couple of calls to free() in the early-return error conditions.

Patch-set: 4
Label: Code-Review=-1