Update patch set 1
Patch Set 1: Code-Review-2

We cannot do this.  This include file is part of the NDK API.
So what we have to do instead is declare two interfaces.

libutils/Looper.h should just contain the pure C++ implementation.  We need to make Looper no longer be a subclass of ALooper and no longer use any types we declared in looper.h.  So we'll have to copy those typedefs and enums to Looper.h and give them new names.

native/.../looper.h should just contain the NDK declaration, basically as-is.

Now the tricky part is that we're duplicating the interface but they could easily get out of sync.  So we'll want to leave comments in Looper.h to the effect that certain declarations (particularly the enums) must match the NDK API definitions.

native/.../looper.c is also going to have to do a little more work to translate between these interfaces.  Should mostly come down to a few casts.

Good luck.  :)

Patch-set: 1
Reviewer: Gerrit User 1003957 <1003957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2