Update patch set 1
Patch Set 1:

After some internal discussion, we're considering just switching the whole binder ABI to be 64 bit safe.  Use uint64_t everywhere we need a pointer (maybe typedef'd to binder_ptr), including the ioctl API and the flattened binder objects.  The upstream kernel would only support the new 64-bit ABI.  In the android-3.10 tree we would switch to the 64-bit ABI, and then add a patch on top with a default-on config option to switch it back to the 32-bit ABI.  libbinder would use a config flag to select between 32-bit and 64-bit, defaulting to the 32-bit for at least the next release, and then changing the default to 64-bit once the new kernels are ubiquitous.  A BoardConfig.mk variable would override the default.

The disadvantage is changing the ABI, but the advantage is that long-term the amount of compatibility code goes to zero.

Patch-set: 1