Fix meta_data_event index error in KeyedVector in SensorService.
SensorService record last sensor data event for non-continous sensor, and this
is using KeyedVector and use 'handle' as the key for each sensor. But meta_data_event
is a special event which shows flush success finished and not belong to any sensor,
so its 'handle' is set to zero(0) due to the requirement, and the key zero is not in the
KeyedVector, so it will cause crash when try to do some action using this key.
And also meta_data_event should not recorded.

Change-Id: I6343cf572f499b86e945a5a676daaffc5bc76b0c
Signed-off-by: Han, He <he.han@intel.com>
Signed-off-by: Wang, Zhifeng <zhifeng.wang@intel.com>