Update patch set 1
Patch Set 1:

One meta-comment:  with our current restorecon logic, this will only run once per unique file_contexts file.  So for example, on an upgrade from 4.4.2, if we have both this change and the change that permits seinfo assignment for third party apps applied, then we'll get the desired behavior on that upgrade.  But this won't address any future updates to the seinfo assignment logic in SELinuxMMAC.java or changes to seapp_contexts that affect the existing seinfo assignments unless there just happens to be a file_contexts change too.

Patch-set: 1