Update patch set 2
Patch Set 2:

The stated goal of the change is to "make struct Event have consistent memory layout".  It seems that the "packed" attribute achieves that goal.  Does it not?

Does the x86_64 compiler generate code that will fail if we use "packed"?  If so, we need to examine any use of "packed" in Android.

Patch-set: 2