Update patch set 1
Patch Set 1:

This is another fall-out from not doing a full recursive restorecon of /data/data and /data/user anymore, but rather doing it per-package.  Alternative would be to try to have libselinux handle this as well as the top-level /data/user directory during the init.rc restorecon_recursive /data.

Patch-set: 1
Label: Verified=0