Update patch set 4
Patch Set 4: Code-Review-2

I don't like the compiler-filter command line option (wtf is a filter) and baking it into frameworks make it harder for me to rename/remove it. The motivation for this change is that if we're stupid we can fix things. I'd rather we worked on understanding and not being stupid rather than providing kludges.

Patch-set: 4
Reviewer: Gerrit User 1010118 <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2