Update patch set 2
Patch Set 2:

we had previously considered the numbered enumeration, but this is just supposed to be for quick debug, we don't want to encourage its use for production.

my point is it is a lot easier to conditionally override dalvik.vm.dex2oat-compiler-filter as a property, than to search through a list of of props to override things.

I'm happy for this fix to go in, but it will be a week or two before I get to moving the compiler filter options to their own more specific properties.

Patch-set: 2