Binder: Resolve race condition
The transaction buffer does not be freed before sendReply,
which may result in the next command fails. e.g. If one
application immediately creates another player to play next
video clip after releases the previous one, the setup may
fail for the native player does not be freed and the surface
is still connected to the previous one.

Change-Id: I2218205232a2878e6099b5e7e4af168481f0ebbf
Signed-off-by: Yiming Cao <ym.caoyiming@gmail.com>