Performance improvement for 2D for large screen devices
On large screen devices using the full backbuffer cause
performance degradation. Hence we propse to use only the dirty region.
Given this is memory intensive operation we suggest to do it only
when the dirty region is smaller than half of the screen size. We
have observed avg 40% performance improvement in 2D rendering with
this patch on our devices.

Signed-off-by: Raj Mamadgi <rmamadgi@sta.samsung.com>

Change-Id: I8f689685a38daf2e221c312234c3fbf93021fcfc