Update patch set 1
Patch Set 1:

It's *incorrect* for a C++ compiler to optimize the test out; the language certainly does not provide the strong non-aliasing guarantees that would make the test genuinely redundant.  It is absolutely possible to write C++ that will wind up calling member functions with a null THIS parameter; it's even possible to do it in a way that the compiler cannot detect. 

There's a productive discussion to be had about whether a hard crash *should* be the behavior here, rather than the current explicit soft failure, but that's not the discussion we're having. :)

Patch-set: 1