Update patch set 1
Patch Set 1: Verified-1

This actually prevents the phone from booting.

This might have gone in as protection for unexpected things, but there's now a bunch of code that relies on this behavior.

This will still need to be fixed, but it will be more involved than what is in this patch.

Patch-set: 1
Label: Verified=-1