Update patch set 1
Patch Set 1: Code-Review-2

Okay, so this was contributed without even checking to see if the system runs without it?  We can have a discussion about whether or not this is right, but unless the system boots without you can't make this change.  If you want to get rid of this, you will need to go through the rest of the code-base and modify anything that depends on it to not do so.

Patch-set: 1
Label: Code-Review=-2