Update patch set 1
Patch Set 1:

> Thanks for the clarification. I agree that we should change code
 > and not to depend on (this == NULL) test. Clang's choice to
 > optimize that code away, if done as a local optimization, is really
 > too aggressive to me.

It isn't necessarily just a clang thing to optimize away "if (this == NULL)" checks. Just because gcc doesn't do it today doesn't mean that they can't change this in the future. Would you feel the same way about the body of an "if (false)" being optimized out? If the "this == NULL" path is ever taken, it is actually a sign of an invalid C++ program, so they use this fact to reduce the comparison to "if (false)". The right way to move forward is to remove the undefined behavior (ticking time bombs) from our codebase, and not to just workaround the optimizer.

Patch-set: 1