Add -Wno-error=undefined-internal for clang
The comments in IPCThreadState.cpp intentionally leaves getReturnString
etc. undefined if LOG_NDEBUG is set.
Make sure clang doesn't fail the build because of this. While clang's
warning is valid (gcc doesn't warn about it because it doesn't warn
about internal functions that are thrown away by the optimizer), it is
intentional.

Change-Id: I625a231fae0c867886ea3567598dbe0d8f3aa475
Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>