Surface: Ensure synchronisation of copyBlt
The Surface::lock() function has an optimisation for copying the previously
drawn contents from the frontbuffer to the (current) backbuffer, so that the
client does not have to redraw the whole surface contents. This logic was not
using the sync fence FD from dequeueBuffer() and so was not correctly
synchronised with pending operations on the backbuffer.

Change-Id: Ib44574d50f8bdb4a23078cd8da1c5c512c876320