Rescale external digitizer axes on rotation
When an external digitizer ('orientationAware = false') is connected to
an Android device, its coordinates are scaled appropriately while the
device is in its natural orientation, but those scale factors are not
updated when the device is rotated. This causes two obvious issues: a
region of the viewport is inaccessible (along the under-scaled axis) and
a region of the digitizer is unusable (along the over-scaled axis).

This patch ensures that the scale values (and mOrientedRanges) are
correct for external digitizers regardless of the viewport orientation.

Change-Id: I0a8a8c3f54b2381dbb950e851a883db14e662d74
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>