Update patch set 2
Patch Set 2:

Checking InputDispatcher.h/.cpp I see no other mutation of the queue head/tail or next pointers of queue members, although I would be more comfortable with a debug assert that count is accurate .

It was never causing huge slowdowns, but with 200hz+ touch and atrace active, if the application did not respond well, the queue count could  reach a few hundred.  With atrace on, count is called for each event to trace the queue count.  This was enough to make InputDispatcher use ~3ms/frame of CPU time and show up as a hotspot of running time in systrace at least.

Patch-set: 2
Reviewer: Gerrit User 1014833 <1014833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0