installd: Patch preoptimized files in data
IsDexOptIsNeeded can return that an odex
file in the dalvik-cache has to be patched.
This case is added here where dexopt get a
new parameter, is_self_patchoat, meaning that
the file needs to be patched in place.

Once boot is completed, the .preloaded flag
is removed to enable the pruning of dalvik-cache
in RelocateImage.

Change-Id: I9e5742193a24945549b092a69cc8a136e16a5d8c
Signed-off-by: Guilhem IMBERTON <guilhem.imberton@intel.com>
Signed-off-by: Julien Delayen <julien.delayen@intel.com>