installd: Patch preoptimized files in data
When patching a file, odex in /system is used.
This does not work for odex preoptimized in /data.
If the odex in /system is not found, we check if the
file already exists in /data and patch it.

Once boot is completed, the .preloaded flag
is removed to enable the pruning of dalvik-cache
in RelocateImage.

Change-Id: I9e5742193a24945549b092a69cc8a136e16a5d8c
Signed-off-by: Julien Delayen <julien.delayen@intel.com>
Signed-off-by: Guilhem IMBERTON <guilhem.imberton@intel.com>