Update patch set 1
Patch Set 1:

Let's go ahead and submit this.  If the code's in our tree it should compile.

I'll open an internal bug to decide what to do with all the tests here.  This specific test might still be useful for some people, but it should probably live under system/extras/tests/framebuffer since it's really tied to the fbdev API and not surfaceflinger.

Patch-set: 1
Label: Code-Review=0