Update patch set 3
Patch Set 3:

> 4) I don't think installd is the place where this task is
 > performed. The framework and art decides whether a compilation or
 > patching is needed and then the installd acts upon that decision.
 > If the content in /data is invalid, dalvik-cache will be erased by
 > art. Also, IsDexOptNeededInternal only return that patching is
 > needed if the classes.dex is valid.

What Richard was asking:

What happens if your boot image is still valid, i.e., ART/zygote is not gonna clean the cache, but your app oat is out of date? Then the PackageManager / isDexOptNeeded will make the decision that it needs to be patched/recompiled (hopefully). Who is cleaning the file? Because if it's not done before installd, then you're just gonna try to patch the old file.

I believe this is done in the PackageManagerService, but I defer to calin and narayan.

Patch-set: 3
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0