Update patch set 1
Patch Set 1:

I've wanted to this for a while as well. 
Last time I looked at it, it was more than just bring-up that needed it. Recovery (and possibly the first boot on encrypted devices -- not sure) used pixelflinger for graphics.

Is there a way we could add restrictions in the build-system, such that new dependencies couldn't be added?

Patch-set: 1