BufferQueue: Prevent possible fence FD leak
If an application doesn't release a native window object,
BufferQueue object is also not released in surfaceflinger
process.

This can lead to fence FD leak because BufferQueue always
has the fence object from the last queueBuffer and
it's not cleared even though the producer or consumer
is disconnected.

So mLastQueueBufferFence should be cleared
when disconnecting producer or consumer to prevent
possible file descriptor leak.

Change-Id: I8c5784081203e7bf14708f0946434e38747e3738