Update patch set 1
Patch Set 1:

> But actually, when producer or consumer is disconnected, all queued
 > buffers are deleted by freeAllBuffersLocked().

Not exactly. When a producer disconnects, every entry in mSlots is cleared. However, the items in mQueue still hold their own references to the GraphicBuffers (see BufferItem), so those buffers will not be deleted until either the BufferQueue is destroyed (taking mQueue with it) or they are acquired by the consumer.

Patch-set: 1
Reviewer: Gerrit User 1043625 <1043625@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0