Update patch set 1
Patch Set 1:

I'm afraid I can't share the series of steps in the application process because I cannot see the application sources. I found this problem when I was repeating launching and pausing the specific internal application on our devices and I just debugged this on the framework side.

As you know, an application can acquire ANativeWindow object when creating a window. ANativeWindow object is actually Surface object, which contains IGraphicBufferProducer object (remote binder object). Even if the window and the related layer is destroyed, BufferQueue object is alive until the application release Surface object. When the problem was occured, the application didn't release their ANativeWindow object while pausing. However, because of this application problem, the number of FD on the surfaceflinger process was increased and, finally, the surfaceflinger process was crashed.

Actually, the problem was on the application side and it has been resolved by changing the application. However, this patch will make the framework more robust.

Patch-set: 1
Reviewer: Gerrit User 1017028 <1017028@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0