Fix the ColorInversion issue if connected to HDMI panel
With color inversion activated, the output on the build-in display
will be corruption if connected to HDMI with different resolution.

Reproduce steps:
1. Boot device connected with HDMI as an external screen.
2. Set HDMI mode to make sure it is different from the build-in display
   resolution, e.g. set HDMI to be 800x600, if the build-in display
   resolution is 1920x1200
3. Activate Color inversion: Settings->Accessibility->Color inversion->On

After step 3), build-in display will be corruption and only 800x600 region
(parts of 1920x1200) is updated.

Root cause: Not updated viewport(display size) before calling
GLES20RenderEngine::beginGroup() to set colorMatrix in
SurfaceFlinger::doDisplayComposition(). So HDMI resolution(800x600) is used
during setting colorMatrix for the built-in display(1920x1200), and only
partial region(800x600) in build-in display is refreshed.

Change-Id: Ia26b7580d58a50aa3256b60f408e0687ca3cb944
Signed-off-by: Yong Yao <yong.yao@intel.com>