Add new method to disable screen capture
In previous code logic, the screen capture is denied if the display
contains secure layer, but setScreenCaptureDisabled just changes
the secure flag of new created Layers, the secure flag of Layers
which is created before calling setScreenCaptureDisabled can not
be updated. Thus the secure flag of all Layers can not been updated
by setScreenCaptureDisabled, SurfaceFlinger can not get the expected
state. We add a flag in DisplayDevice to control the screen capture,
when setScreenCaptureDisabled set the enable/disable state, the
state is updated to DisplayDevice immediately, SurfaceFlinger can
get the expected state when doing screen capture. This fix has two
part, here is the native part.

Change-Id: Ia5ae3d7f2af2d4e20ad51e87e3dffc6f528f9bc3