Update patch set 1
Patch Set 1:

> This is not necessary.  The condition is "i-- > 0" which means we
 > are comparing the old value with 0 before decrementing it.
 > 
 > Although overflow occurs, after the loop is done we really don't
 > care what value i has.

Yes, I agree that the value is not used, and the comparison does in fact take place before the decrement occurs. However, when enabling fsanitize=integer, which performs unsigned and signed integer overflow checks and will be enabled in a future cl, the condition triggers an unsigned overflow.

Patch-set: 1
Label: Verified=0