Update patch set 1
Patch Set 1:

It will be common for applications to have both libGLESv2 and libGLESv3 in the same process. They both export the same symbols. Which one gets used? Not sure the answer matters in practice..

Is there a way to create a "proxy" .so? I think this means a .so that exports symbols but doesn't define them, but pulls in definitions from a dependent library. So if you link against libGLESv3.so, at runtime the linker will load it, load the libGLESv2.so it depends on, and when trying to resolve symbols in the app it first looks in libGLESv3.so but if they're not found it looks in libGLESv2.so and finds them there.

Patch-set: 1