Update patch set 1
Patch Set 1:

Circling back around on this after a conversation with sadmac@ this morning.

I think we agree that the small amount of code necessary to implement both versions e.g.

  status_t Parcel::readByteVector(
      std::vector<uint8_t>* out);
  status_t Parcel::readByteVector(
      std::unique_ptr<std::vector<uint8_t>>* out);

is worth it for the better semantics around interface generation for native code: if you handle nulls, you get a pointer, if you don't, you get a value reference.  By changing signatures when nulls are allowed or not, you cannot get yourself into a situation where the code expects nulls (or not) and the generated code thinks something different.

Patch-set: 1
Label: Code-Review=0