Update patch set 1
Patch Set 1: Presubmit-Ready+1

I still need to verify that this doesn't break mips/mips64 tomorrow (on the emulator), but I believe this fixes all of the issues. I am not sure if there are other tests to run (other than just booting the device, which was what fails when we don't have the assembly constraints correct here).

Patch-set: 1
Label: Presubmit-Ready=0