Update patch set 1
Patch Set 1:

It's possible there are some bugs around this, and I think we may be changing some of the logic around Layer destruction in an upcoming release, but this change in particular doesn't seem to be correct to me. Do you see a stack coming into ConsumerBase::abandonLocked where the lock isn't held?

Patch-set: 1