Fix memory ordering issues; document IMemory peculiarities
Convert to standard atomics.

Correct mHeapId memory ordering. Required acquire ordering on loads
was missing in several places.

Remove atomic updates to count, since it is only updated with lock
held. (And would be missing fences if this were not true.)

Document the peculiar use of copy-in-write vectors in a context in
which copy-on-write is unsafe.

While we're here, fix an extremely infrequent file descriptor
leak discovered by Yabin, and consistently check dup() for errors.

Note that the file descriptor leak was already fixed differently
in internal trees.

Merged-in: I05b9f96e3867fa2e0abe6f319be8c56b89624c41

Bug: 28816986

Change-Id: I05b9f96e3867fa2e0abe6f319be8c56b89624c41