Update patch set 2
Patch Set 2:

The change doesn't seem wrong to me, but the motivation is suspect.

Why is libpixelflinger.so linking against libEGL.so? That violates the layering of the system, dependencies shouldn't go that way. I'm also not seeing that dependency.

This loop only arises if an EGL driver is linking against libEGL.so, which they should not be doing. So I'm inclined not to fix it.

Patch-set: 2
Reviewer: Gerrit User 1010672 <1010672@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0