Update patch set 1
Patch Set 1: Verified+1 Code-Review+2

I don't know what the problem with window format is, but in the current code we'll first disconnect the window, and then call the driver's eglReleaseThread. If the driver has a buffer dequeued from the current surface, it will try to cancel that buffer, and that will fail because the window is disconnected. I don't think that's a real problem in practice, but it's ugly and likely causes an error log.

Patch-set: 1
Reviewer: Gerrit User 1010672 <1010672@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=+1