Update patch set 1
Patch Set 1:

Our current code does, for example,

        case SET_AUTO_REFRESH: {
            CHECK_INTERFACE(IGraphicBuffer, data, reply);
            bool autoRefresh = data.readInt32();
            status_t result = setAutoRefresh(autoRefresh);
            reply->writeInt32(result);
            return NO_ERROR;
        }

The service error is encoded as a part of the reply, rather than resulting a transaction error (and we do not really generate any transaction error like we should).

When converting to binder::Status, we want to continue generating EX_SERVICE_SPECIFIC errors that use status_t as the error code.

Patch-set: 1