Update patch set 1
Patch Set 1:

Thanks for the feedback, Jesse.  I've passed it along to the author.

We'll have to do a fair amount of testing on any changes to that area of the code and it may be a while before we get our hands down in there to make your suggested revisions.  So we're not ignoring this, just waiting for an opportunity to get back in that part of the meta-EGL library.

Patch-set: 1
Reviewer: Gerrit User 1099483 <1099483@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0