Update patch set 2
Patch Set 2:

(3 comments)

Thanks!

Do you guys believe my reasoning that count does not need to be atomic? That seems to be the highest risk part of this, if we want to take it for MR1. But leaving it alone for now seems like a worse option.

Patch-set: 2
Label: Verified=0