inputflinger: make Touch work for Mosaic display mode
This is to support new Touch feature for Mosaic display mode, which
means Android UX content is spanned across both 2 displays as if they
were a single display. In this mode, Touch input from each display will
only reflected on the content of its own display.

In this patch, two configurable parameters are introduced in idc file,
  touch.mosaic.left and touch.mosaic.right
the value of these two parameter can be "true" or the phys of its input
device, which is used when two same touch controller are applied.

Based on Mosaic left/right setting, surface's left and width are adjusted
to do right mapping.

Test: work well on Mosaic display mode
Change-Id: I85333a7cd5ec4b7c0699e13aff05f9032e08ad26
Signed-off-by: Qipeng Zha <qipeng.zha@intel.com>
Signed-off-by: Liu Zhiquan <zhiquan.liu@intel.com>