Fix EGLImage creation when protected content is requested
In aec8697f support for EXT_protected_surface was added. However, this
will not work correctly if the system doesn't support
EGL_ANDROID_image_crop or the crop rectangle isn't valid. In this case
the third entry of the attribute list will be set to EGL_NONE and the
rest of the list will be ignored by the graphics driver.

Fix this by dynamically setting EGL_PROTECTED_CONTENT_EXT to the correct
position.

Test: Create an EGLImage from a window surface that has the
      GRALLOC_USAGE_PROTECTED usage bit set

Change-Id: I7829f3a627814726f828639aa6630dd9a8711fae
Signed-off-by: Christian Poetzsch <christian.potzsch@imgtec.com>