Update patch set 1
Patch Set 1:

(1 comment)

> (3 comments)
 > 
 > Is there a bug for this?
 > 
 > Also, these seem similar to system/libhwbinder/tests/benchmarks.
 > Should we delete those?

I think we should keep those, as they use functionality that libbinder doesn't support (eg scatter-gather). Ideally we have a single test that covers the intersection of libbinder and libhwbinder for both (that can be this one), and then a set of extended tests just for libhwbinder's extended functionality.

Patch-set: 1
Label: Verified=0