Update patch set 1
Patch Set 1:

> Actually, syspropsChanged seems broken somehow.  You want to state
 > which system property has changed.

Every Android app that loads libcutils can register a callback to be called when system properties change. This callback is not tied to any particular system property. Current users of this are atrace (which simply reloads the relevant trace properties when it gets the callback), and the Java framework, which has SystemProperties.addChangeCallback(), and clients like WindowManager.

So while we can certainly make this method take a property or a list of properties that have changed, the underlying infrastructure won't use it. Maybe we can change how libcutils / the framework deal with this, but for now this isn't the best thing to spend our time on I think.

Patch-set: 1
Label: Verified=0