Sensor: Add ASensorEventQueue_unregisterSensor() function
Currently, when sensor is activated using ASensorEventQueue_registerSensor
function, it is deactivated using ASensorEventQueue_disableSensor function.
With respect to ASensorEventQueue_registerSensor function,
ASensorEventQueue_disableSensor does not match to be its corresponding
function for deactivating the sensor. Therefore, lets add
ASensorEventQueue_unregisterSensor function to deactivate the sensor for a
better matching with the corresponding sensor activation function.

Test: In a ndk based application, activate the sensor using
ASensorEventQueue_registerSensor and deactivate the sensor using
ASensorEventQueue_unregisterSensor. Check if sensor is deactivated or not.

Change-Id: Id38c002b57154c1feedf099f45b0ed8ed004ca39
Signed-off-by: Aniroop Mathur <a.mathur@samsung.com>