Update patch set 1
Patch Set 1:

Over in https://android-review.googlesource.com/#/c/313169/ I wrote tests for the GID ranges, and I realized that the shared GIDs were actually "leaking" outside of their defined range.

The builder method over there now returns -1 when an invalid appId is passed in, so this CL is a band-aid to avoid erroring out.  If the profile stuff really needs to work for non-apps, then we should clean up the GID ranges to be more explicit about how they behave.

Patch-set: 1