Update patch set 1
Patch Set 1:

I replied to the email, but this will basically trigger a bunch of warnings in the profile handling on the art side. We protect ourselves so there will be no hard crash, but it's an unexpected state to not have profiles when dalvik.vm.usejitprofiles is set.

Let's try to address this. We can continue in the mail thread.

Patch-set: 1
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0