Update patch set 1
Patch Set 1:

dimitry and I had a talk about this, and it sounds like we're probably just going to drop versioning for the NDK for anything that isn't versioned in the platform libraries (basically everything but bionic) for the time being.

It should be safe, but there is the risk that if we push these versioned stubs out in the NDK and then the platform starts versioning the symbols but with a different version name, we'll break a bunch of apps.

Still, we're going to need to get some docs together about how symbol versioning works in the platform (dimitry is going to start on that, and I'll add the NDK relevant pieces).

Patch-set: 1
Label: Autosubmit=0