Update patch set 2
Patch Set 2:

(1 comment)

Why not have a main method that parses the arguments/etc... and passes them to lshal? I'm okay with the split, but I think the argument parsing should be decoupled from the lshal class which is what inspects the system.

LsHal? lshal? Lshal? ¯\_(ツ)_/¯ Lshal.

Patch-set: 2
Label: Verified=0