Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

Looks like an issue that we crash dex2oat but this is probably not the right solution.

We should still get this as an optimization (as we shouldn't bother open the app image otherwise) but dex2oat should be fixed not to crash.

Patch-set: 3
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1