HACK: test-hwc2: remove need to recompile system.img
This is a temporary work around.

Previously, to run the hwc2 tests,
"#define HAVE_NO_SURFACE_FLINGER" had to be set in
BufferQueueCore.cpp and recompiled. This change informs
BufferQueueCore.cpp at runtime that surfaceflinger is not running.
With this change, #define HAVE_NO_SURFACE_FLINGER does not have
to be set so the system.img does not need to be recompiled.

If the test cases are stopped at the wrong time, the system will
be in an bad state. This does not affect subsequent attempts to
run the test cases. It becomes a problem when SurfaceFlinger is
restarted. On some devices, it is not obvious that the system
is in an bad state. To restore the system to a valid state,
the system property config.headless must be manually set to 0 or
the device must be rebooted.

For these reasons, DO NOT MERGE THIS CHANGE!

Future changes are expected to make this work around unnecessary.

Test: run the test-hwc2 tests

Change-Id: I9d18d205ce3c8b89603081a14f085236cba4407e