Fix docs build to use the just built NDK.
INPUT is now passed from the command line so we can pass it
$SOONG_OUT (still a TODO to actually use SOONG_OUT, because that's
not accessible outside the build system).

Drop the "website" directory, since that doesn't appear to be used.

Add a clean target, and add the output to the gitignore.

This still isn't great (it needs to be integrated into the docs
target), but it corrects the current situation without making things
any worse.

Test: make -C frameworks/native/docs
Bug: None
Change-Id: Idc2db0f90f7084f72c8591b3e4ad5fdfd6ddb47e