Override event timestamp with uint32_t MSC_TIMESTAMP if available.
Some hardware devices (e.g. DualShock4 motion sensors) send out the
hardware timestamp along with the events. For time-sensitive data,
this can be very important, because the default event timestamp,
which is the system time when the event enters the system, can vary
largely from the actual time that the event was generated by the
hardware.

The MSC_TIMESTAMP is used to report such hardware timestamp in the
kernel, and as such, we should override the event timestamp whenever
this timestamp is available. Since MSC_TIMESTAMP initial state is
different from the system time, we calculate the difference between
two MSC_TIMESTAMP events, and use that to calculate the system time
that should be tagged on the event.

Test: No stand-alone test.
Change-Id: I06833d3b517ed0db5fc5ae02df84ca5831b27e1e