Added INPUT_DEVICE_CLASS_ACCELEROMETER, which uses JoystickInputMapper
A separate device class is needed to identify an accelerometer device,
which may contain a gyroscope as well. This is done by checking the
INPUT_PROP_ACCELEROMETER bit in the kernel device property mask.

Accelerometer/gyroscope device will also be classified as joysticks,
because it has similar axis as a joystick, and as such, can reuse
the joystick mapper. JoystickInputMapper now has two sources, a
regular joystick (AINPUT_SOURCE_JOYSTICK) or an accelerometer
(AINPUT_SOURCE_ACCELEROMETER).

Test: Connect a DualShock4, and run "adb shell dumpsys input".
      Check that there is a JoystickInputMapper under the motion
      sensors device.
Change-Id: Iaae06a4bc58b6705fe01079bd06d0f84baf49daf