Update patch set 2
Patch Set 2:

The modification in layer works for me (it's the same as we made before).
In this fix, layer didn't get the opacity of sideband handle, I think it is not good enough. 
Surfaceflinger can't get the opacity of a handle while HWC knows, that's why I think surfaceflinger need just to clear as HWC asked.

Patch-set: 2
Reviewer: Gerrit User 1157249 <1157249@85c56323-6fa9-3386-8a01-6480fb634889>