Update patch set 5
Patch Set 5:

> (1 comment)

This is a contradictory thing.
We want to ensure that the loop is run at least once.
But also to avoid chainIndex-- to produce overflow.
So we do not use chainIndex as a loop condition.
Break loop only when the !chainIndex.

Patch-set: 5
Reviewer: Gerrit User 1106243 <1106243@85c56323-6fa9-3386-8a01-6480fb634889>