dumpstate: use tombstoned/debuggerd for java traces.
- debuggerd_dump_traces now supports Java traces as well, so
  use that when dalvik.vm.stack-trace-dir is set. The output
  FD we use for the intercept is currently a regular file
  written by mkostemp, but this can be cleaned up once the
  old way of doing things is removed.

- We're no longer writing traces to a global trace file, so
  add bug report entries for all java traces written in the past
  30 minutes just as we do for native traces.

The minor refactoring that this CL undertakes was done under the
assumption that the old way of doing things will be removed in the
near future.

TODO in a followup PS: the unlink() of the temporary file created
fails because it's owned by root. This will be fixed in a future
PS.

Bug: 32064548
Test: manual

Change-Id: I8af6f0a644115296dc41affc3b7cd98a2db32c48